+++ submit-api/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 3.0.0 -- October 2020
+
+* Upgrade dependencies to more recent version of the Cardano eco-system (cardano-db-sync-6.0.0) 
+* Interface **breaking change**: the command-line no longer accept a `--security-param` option.
+
 ## 2.1.3 -- August 2020
 
  * No changes. Bumping version to keep consistency with explorer-api.